requirements.txt
# Note: CMake must be installed first:
# macOS: brew install cmake
# Ubuntu: apt-get install cmake
# Windows: Download from cmake.org

flask>=2.0.0
requests>=2.28.0
pandas>=1.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
uvicorn>=0.15.0
websockets>=10.0
yolov5>=7.0.0
fastapi>=0.70.0
opencv-python-headless>=4.5.0
mediapipe>=0.8.0
torch>=1.10.0
numpy>=1.21.0
mysql-connector-python>=8.0.0
SQLAlchemy>=1.4.0
python-multipart>=0.0.5
face-recognition>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
